Move annotation classes into an annotation package.
