+++ src/main/java/htsjdk/beta/codecs/variants/vcf/VCFCodec.java
@@ -9,7 +9,7 @@ import htsjdk.beta.plugin.variants.VariantsFormats;
 import htsjdk.samtools.util.BlockCompressedStreamConstants;
 import htsjdk.samtools.util.FileExtensions;
 import htsjdk.samtools.util.IOUtil;
-import htsjdk.utils.InternalAPI;
+import htsjdk.annotations.InternalAPI;
 import htsjdk.utils.ValidationUtils;
 
 import java.io.IOException;
