+++ src/main/java/htsjdk/beta/package-info.java
@@ -1,12 +1,12 @@
 @BetaAPI
 package htsjdk.beta;
 
-import htsjdk.utils.BetaAPI;
+import htsjdk.annotations.BetaAPI;
 
 /**
  * This package is a staging area used to temporarily house new packages that are released as
- * for evaluation as part of the {@link htsjdk.utils.BetaAPI}. The entire contents of this package is
- * {@link htsjdk.utils.BetaAPI}. Subpackages may be subsequently released as part of the public API,
+ * for evaluation as part of the {@link htsjdk.annotations.BetaAPI}. The entire contents of this package is
+ * {@link htsjdk.annotations.BetaAPI}. Subpackages may be subsequently released as part of the public API,
  * at which time they will be moved out of the {@link htsjdk.beta} package.
  */
