+++ src/main/java/htsjdk/beta/plugin/interval/HtsIntervalUtils.java
@@ -3,7 +3,7 @@ package htsjdk.beta.plugin.interval;
 import htsjdk.samtools.QueryInterval;
 import htsjdk.samtools.SAMSequenceDictionary;
 import htsjdk.samtools.util.Locatable;
-import htsjdk.utils.InternalAPI;
+import htsjdk.annotations.InternalAPI;
 import htsjdk.utils.ValidationUtils;
 
 import java.util.Arrays;
