+++ src/main/java/htsjdk/beta/plugin/package-info.java
@@ -1,9 +1,9 @@
 @BetaAPI
 package htsjdk.beta.plugin;
 
-import htsjdk.utils.BetaAPI;
+import htsjdk.annotations.BetaAPI;
 
 /**
- *  The entire contents of this package is {@link htsjdk.utils.BetaAPI}. Code and subpackages may be subsequently
+ *  The entire contents of this package is {@link htsjdk.annotations.BetaAPI}. Code and subpackages may be subsequently
  *  released as part of the public API, at which time they will be moved out of the {@link htsjdk.beta} package.
  */
